+++ src/Infrastructure/Laravel/database/migrations/2022_02_21_144644_create_basic_trader_tables.php
@@ -35,6 +35,7 @@ return new class extends Migration {
             $table->char('variant_id', 36)->primary();
             $table->char('product_id', 36);
             $table->string('sku')->unique();
+            $table->string('ean')->unique()->nullable();
             $table->boolean('show_in_grid')->default(0);
             $table->string('state')->default(\Thinktomorrow\Trader\Domain\Model\Product\Variant\VariantState::available->value);
             $table->integer('sale_price')->unsigned();